Add files via upload
The purpose of this assignment is to test the MittArv application on a mobile device, identify
critical UI/UX issues, and suggestimprovements. At MittArv it's essentialthatthe user experience
is smooth, intuitive, and free from major issues.
The purpose of this assignment is to validate the MittArv postlogin website, and develop a
sample suite of automation scripts in the language of your choice.